+++ routes/pagination.go
@@ -1,9 +1,36 @@
 package routes
 
+import (
+	"gorm.io/gorm"
+	"net/http"
+	"strconv"
+)
+
 type PaginatedResponse struct {
 	NumPages  uint
 	Page      uint
 	PageSize  uint
 	TotalSite uint
 	LastPage  bool
+}
+
+func Paginate(r *http.Request) func(db *gorm.DB) *gorm.DB {
+	return func(db *gorm.DB) *gorm.DB {
+		q := r.URL.Query()
+		page, _ := strconv.Atoi(q.Get("page"))
+		if page == 0 {
+			page = 1
+		}
+
+		pageSize, _ := strconv.Atoi(q.Get("page_size"))
+		switch {
+		case pageSize > 100:
+			pageSize = 100
+		case pageSize <= 0:
+			pageSize = 10
+		}
+
+		offset := (page - 1) * pageSize
+		return db.Offset(offset).Limit(pageSize)
+	}
 }
